+++ scripts/brazilScripts/BrazilLandmarks/Landmarks.js
@@ -1,0 +1,5 @@
+export function landmarkCard(landmarkObject){
+    return`
+    <li class = "list-item">${landmarkObject.name}</li>`
+
+}
