+++ scripts/brazilScripts/BrazilLandmarks/landmarksDataProvider.js
@@ -1,0 +1,19 @@
+let brazilLandmarks = [
+    
+    {
+        name:  "Iguazu Falls"
+    },
+
+    {
+        name:  "Christ the Redeemer"
+    },
+   
+    {
+        name:  "Sugar Loaf Mountain"
+    }
+    
+]
+
+export function useLandmarks(){
+    return brazilLandmarks.slice()
+}
