+++ scripts/brazilScripts/main.js
@@ -5,4 +5,7 @@ cityList()
 import { peopleList } from "./BrazilFamousPeople/famousPeopleList.js"
 peopleList()
 
+import { landmarksList } from "./BrazilLandmarks/landmarksList.js"
+landmarksList()
+
